[Smoke-Fails] Adds ninja check-openmp as smoke-fails test
This requires the ninja build system to be used!
The added test acts as a proxy to run ninja check-openmp in the
AOMP_REPOS/build/openmp build directory.

I am unsure how to make it easily switchable from ninja to make or the
other way around, as this is not exported in the environment. One could
probably build a whole lot of magic to detect it, but for now.. well.
Let's see if that works for people in the first place, when they build
AOMP in the regular location and with ninja.