[BBots] Add upstream ubu-22 Dockerfile and README
To help developers, we containerize more of our upstream testing and
make the Dockerfiles publicly available. This is an effort to help to
enable developers to reproduce problems on the buildbots locally.

This adds the first Dockerfile for a Ubuntu 22 based dockerized
post-commit buildbot that uses the AMDGPUBot.cmake CMake cache file
inside the llvm-project/offload/cmake/cache directory.